+++ Demo/DemoTests/MoyaError+MoyaSpec.swift
@@ -1,50 +1,51 @@
 import Moya
 
-extension MoyaError {
-    public func sameErrorType(otherError: MoyaError) -> Bool {
-        switch self {
+extension MoyaError: Equatable { }
+
+public func ==(lhs: MoyaError, rhs: MoyaError) -> Bool {
+    
+    switch lhs {
+    case .ImageMapping:
+        switch rhs {
         case .ImageMapping:
-            switch otherError {
-            case .ImageMapping:
-                return true
-            default:
-                return false
-            }
+            return true
+        default:
+            return false
+        }
+    case .JSONMapping:
+        switch rhs {
         case .JSONMapping:
-            switch otherError {
-            case .JSONMapping:
-                return true
-            default:
-                return false
-            }
+            return true
+        default:
+            return false
+        }
+    case .StringMapping:
+        switch rhs {
         case .StringMapping:
-            switch otherError {
-            case .StringMapping:
-                return true
-            default:
-                return false
-            }
+            return true
+        default:
+            return false
+        }
+    case .StatusCode:
+        switch rhs {
         case .StatusCode:
-            switch otherError {
-            case .StatusCode:
-                return true
-            default:
-                return false
-            }
+            return true
+        default:
+            return false
+        }
+    case .Data:
+        switch rhs {
         case .Data:
-            switch otherError {
-            case .Data:
-                return true
-            default:
-                return false
-            }
+            return true
+        default:
+            return false
+        }
+    case .Underlying:
+        switch rhs {
         case .Underlying:
-            switch otherError {
-            case .Underlying:
-                return true
-            default:
-                return false
-            }
+            return true
+        default:
+            return false
         }
     }
 }